+++ L2J_ReunionProject_Core/java/l2r/gameserver/model/actor/stat/PcStat.java
@@ -608,7 +608,7 @@ public class PcStat extends PlayableStat
 	 * @return the base move speed of given movement type.
 	 */
 	@Override
-	public float getBaseMoveSpeed(MoveType type)
+	public double getBaseMoveSpeed(MoveType type)
 	{
 		final L2PcInstance player = getActiveChar();
 		if (player.isTransformed())
@@ -631,9 +631,9 @@ public class PcStat extends PlayableStat
 	}
 	
 	@Override
-	public int getRunSpeed()
+	public double getRunSpeed()
 	{
-		int val = super.getRunSpeed() + Config.RUN_SPD_BOOST;
+		double val = super.getRunSpeed() + Config.RUN_SPD_BOOST;
 		
 		// Apply max run speed cap.
 		if ((val > FormulasConfigs.MAX_RUN_SPEED) && !getActiveChar().canOverrideCond(PcCondOverride.MAX_STATS_VALUE))
@@ -660,9 +660,9 @@ public class PcStat extends PlayableStat
 	}
 	
 	@Override
-	public int getWalkSpeed()
+	public double getWalkSpeed()
 	{
-		int val = super.getWalkSpeed() + Config.RUN_SPD_BOOST;
+		double val = super.getWalkSpeed() + Config.RUN_SPD_BOOST;
 		
 		// Apply max run speed cap.
 		if ((val > FormulasConfigs.MAX_RUN_SPEED) && !getActiveChar().canOverrideCond(PcCondOverride.MAX_STATS_VALUE))
